Exception handling - delete

+++ src/main/java/com/said/apps/services/UserService.java
@@ -4,10 +4,12 @@ import java.util.List;
 import java.util.Optional;
 
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.dao.DataIntegrityViolationException;
 import org.springframework.stereotype.Service;
 
 import com.said.apps.entities.User;
 import com.said.apps.repositories.UserRepository;
+import com.said.apps.services.exceptions.DatabaseException;
 import com.said.apps.services.exceptions.ResourceNotFoundException;
 
 @Service
@@ -30,16 +32,24 @@ public class UserService {
 	}
 	
 	public void delete(Long id) {
-		repository.deleteById(id);
+		try {
+			if (!repository.existsById(id)) {
+				throw new ResourceNotFoundException(id);
+			}
+			repository.deleteById(id);
+		}
+		catch (DataIntegrityViolationException e) {
+			throw new DatabaseException(e.getMessage());
+		}
 	}
 	
 	public User update(Long id, User obj) {
 		User entity = repository.getReferenceById(id);
-		updateDate(entity, obj);
+		updateData(entity, obj);
 		return repository.save(entity);
 	}
 
-	private void updateDate(User entity, User obj) {
+	private void updateData(User entity, User obj) {
 		entity.setName(obj.getName());
 		entity.setEmail(obj.getEmail());
 		entity.setPhone(obj.getPhone());
